structure output better
I made the output into a tabular format by aligning the items using the format string in the printTask function. The format string now has %-5d for the "number" column, %-10s for the "status" column, and %-30s for the "description" column, so they will be left-aligned and have equal space in the output. The duration will follow that making it aligned correctly as well. The "<|end-of-line|>" is used instead of \n to define a new line that is part of the code structure.

Resolves #9