Support subtasks
Added a new command 'add-subtask <tasknumber>' to add a nested task that can be tracked.
When printing tasks, it now prints subtasks under their respective parent tasks.
When working on a subtask designated as '1.1' for example, it will increment timer for the subtask and the total time worked on task 1 will include the time worked on the subtask.
In the handleCommand function, it now handles the 'add-subtask' command and it also handles a 'work' command on a subtask.
In workOnTask function, it now checks if the task number includes a '.', if it does, it splits it into two parts, the part before the dot is the index of a main task and the part after the dot is the index of a subtask.

Resolves #20