Get rid of globals
Added a new type `TaskList` to encapsulate the list of tasks and associated methods. Consequently, removed the global variable `tasks` and updated functions to be methods receiving a `TaskList`. Also `loadTasksFromFile`, `handleCommand`, and `saveTasksToFile` functions now accept the `TaskList` as a parameter. In `main`, an instance of `TaskList` is created and used for operations instead of the global `tasks` variable. Further required changes to other parts of the code have been kept minimal.

Resolves #23